+++ src/grpc/grpc_interceptors.cpp
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include "grpc_interceptors.hpp"
+#include "grpc/grpc_interceptors.hpp"
 
 #include <grpcpp/support/interceptor.h>